+++ packages/webgal/src/Core/gameScripts/setVar.ts
@@ -4,11 +4,10 @@ import { webgalStore } from '@/store/store';
 import { setStageVar } from '@/store/stageReducer';
 import { logger } from '@/Core/util/logger';
 import { compile } from 'angular-expressions';
-import { setConfigData, setGlobalVar } from '@/store/userDataReducer';
+import { setGlobalVar } from '@/store/userDataReducer';
 import { ActionCreatorWithPayload } from '@reduxjs/toolkit';
 import { ISetGameVar } from '@/store/stageInterface';
 import { dumpToStorageFast } from '@/Core/controller/storage/storageController';
-import { WebGAL } from '@/Core/WebGAL';
 /**
  * 设置变量
  * @param sentence
@@ -21,7 +20,6 @@ export const setVar = (sentence: ISentence): IPerform => {
     }
   });
   let targetReducerFunction: ActionCreatorWithPayload<ISetGameVar, string>;
-  const configData = webgalStore.getState().userData.configData;
   if (setGlobal) {
     targetReducerFunction = setGlobalVar;
   } else {
@@ -60,17 +58,12 @@ export const setVar = (sentence: ISentence): IPerform => {
     } else {
       if (!isNaN(Number(valExp))) {
         webgalStore.dispatch(targetReducerFunction({ key, value: Number(valExp) }));
-      } else if (configData[key]) {
-        webgalStore.dispatch(setConfigData({ key, value: valExp }));
       } else {
         webgalStore.dispatch(targetReducerFunction({ key, value: valExp }));
       }
     }
     if (setGlobal) {
       logger.debug('设置全局变量：', { key, value: webgalStore.getState().userData.globalGameVar[key] });
-      dumpToStorageFast();
-    } else if (configData[key]) {
-      logger.debug('设置配置变量：', { key, value: webgalStore.getState().userData.configData[key] });
       dumpToStorageFast();
     } else {
       logger.debug('设置变量：', { key, value: webgalStore.getState().stage.GameVar[key] });
@@ -92,27 +85,23 @@ type BaseVal = string | number | boolean;
 export function getValueFromState(key: string) {
   let ret: any;
   const stage = webgalStore.getState().stage;
-  const GUI = webgalStore.getState().GUI;
   const userData = webgalStore.getState().userData;
-  const _Merge = { ...GUI, ...stage, ...userData };
+  const _Merge = { ...stage, ...userData };
   const is_baseVal = (_obj: object, _val: any, no_get = false) =>
     ['string', 'number', 'boolean'].includes(typeof (no_get ? _val : Reflect.get(_obj, _val)));
   let _all: { [key: PropertyKey]: any } = {};
-  // 排除GameVar、globalGameVar、configData因为这几个本来就可以获取
+  // 排除GameVar、globalGameVar因为这几个本来就可以获取
   for (let i in _Merge) {
     if (i === 'GameVar') continue;
     if (i === 'globalGameVar') continue;
-    if (i === 'configData') continue;
     if (i) {
-      Reflect.set(_all, '$' + i, Reflect.get(_Merge, i) as BaseVal);
+      _all['$' + i] = Reflect.get(_Merge, i) as BaseVal;
     }
   }
   if (stage.GameVar.hasOwnProperty(key)) {
     ret = stage.GameVar[key];
   } else if (userData.globalGameVar.hasOwnProperty(key)) {
     ret = userData.globalGameVar[key];
-  } else if (is_baseVal(userData.configData, key)) {
-    ret = compile(key)(userData.configData) as BaseVal;
   } else if (key.startsWith('$') && is_baseVal(_all, compile(key)(_all), true)) {
     ret = compile(key)(_all) as BaseVal;
   } else {
