+++ packages/webgal/src/UI/Title/Title.tsx
@@ -6,23 +6,14 @@ import { useDispatch, useSelector } from 'react-redux';
 import { RootState, webgalStore } from '@/store/store';
 import { setMenuPanelTag, setVisibility } from '@/store/GUIReducer';
 import { MenuPanelTag } from '@/store/guiInterface';
-import { nextSentence } from '@/Core/controller/gamePlay/nextSentence';
-import { restorePerform } from '@/Core/controller/storage/jumpFromBacklog';
-import { setEbg } from '@/Core/gameScripts/changeBg/setEbg';
 import useTrans from '@/hooks/useTrans';
 // import { resize } from '@/Core/util/resize';
 import { hasFastSaveRecord, loadFastSaveGame } from '@/Core/controller/storage/fastSaveLoad';
 import useSoundEffect from '@/hooks/useSoundEffect';
-import { WebGAL } from '@/Core/WebGAL';
 import useApplyStyle from '@/hooks/useApplyStyle';
-import { fullScreenOption } from '@/store/userDataInterface';
+import { fullScreenOption, IConfigData } from '@/store/userDataInterface';
 import { keyboard } from '@/hooks/useHotkey';
-import { logger } from '@/Core/util/logger';
-import { assetSetter, fileType } from '@/Core/util/gameAssetsAccess/assetSetter';
-import { setGuiAsset, setLogoImage } from '@/store/GUIReducer';
-import { getStorage, setStorage } from '@/Core/controller/storage/storageController';
-import { getFastSaveFromStorage, getSavesFromStorage } from '@/Core/controller/storage/savesController';
-import { setUserData } from '@/store/userDataReducer';
+import useConfigData from '@/hooks/useConfigData';
 /**
  * 标题页
  * @constructor
@@ -36,51 +27,10 @@ const Title: FC = () => {
   const showBackground = background === '' ? 'rgba(0,0,0,1)' : `url("${background}")`;
   const t = useTrans('title.');
   const { playSeEnter, playSeClick } = useSoundEffect();
+  const configData = useSelector((state: RootState) => state.userData.globalGameVar) as IConfigData;
 
   const applyStyle = useApplyStyle('UI/Title/title.scss');
-  useEffect(() => {
-    // configData发生变化
-    const configData = webgalStore.getState().userData.configData;
-    for (let i in configData) {
-      if ((configData[i] as string) !== WebGAL.ConfigData[i]) {
-        const val = configData[i] as string;
-        switch (i) {
-          case 'Title_img': {
-            const titleUrl = assetSetter(val, fileType.background);
-            webgalStore.dispatch(setGuiAsset({ asset: 'titleBg', value: titleUrl }));
-            setEbg(titleUrl);
-            break;
-          }
-
-          case 'Game_Logo': {
-            const logoUrlList = [assetSetter(val, fileType.background)];
-            webgalStore.dispatch(setLogoImage(logoUrlList));
-            break;
-          }
-
-          case 'Title_bgm': {
-            const bgmUrl = assetSetter(val, fileType.bgm);
-            webgalStore.dispatch(setGuiAsset({ asset: 'titleBgm', value: bgmUrl }));
-            break;
-          }
-
-          case 'Game_name': {
-            WebGAL.gameName = val;
-            document.title = val;
-            break;
-          }
-
-          case 'Game_key': {
-            WebGAL.gameKey = val;
-            getStorage();
-            getFastSaveFromStorage();
-            getSavesFromStorage(0, 0);
-            break;
-          }
-        }
-      }
-    }
-  }, [webgalStore.getState().userData.configData]);
+  useConfigData(configData); // 监听基础ConfigData变化
   return (
     <>
       {GUIState.showTitle && <div className={applyStyle('Title_backup_background', styles.Title_backup_background)} />}
