+++ packages/webgal/src/store/userDataReducer.ts
@@ -8,7 +8,6 @@ import {
   IAppreciationAsset,
   IOptionData,
   ISaveData,
-  ISetConfigDataPayload,
   ISetOptionDataPayload,
   ISetUserDataPayload,
   IUserData,
@@ -47,7 +46,6 @@ export const initState: IUserData = {
     bgm: [],
     cg: [],
   },
-  configData: {}, // 存放在config中定义的全局变量
 };
 
 const userDataSlice = createSlice({
@@ -62,10 +60,6 @@ const userDataSlice = createSlice({
     setUserData: (state, action: PayloadAction<ISetUserDataPayload>) => {
       const { key, value } = action.payload;
       state[key] = value;
-    },
-    setConfigData: (state, action: PayloadAction<ISetConfigDataPayload>) => {
-      const { key, value } = action.payload;
-      state.configData[key] = value;
     },
     unlockCgInUserData: (state, action: PayloadAction<IAppreciationAsset>) => {
       const { name, url, series } = action.payload;
@@ -141,7 +135,6 @@ const userDataSlice = createSlice({
 
 export const {
   setUserData,
-  setConfigData,
   resetUserData,
   setOptionData,
   setGlobalVar,
